renamed the folder back to fusion
