+++ fusion/PanelFooter.js
