+++ fusion/Ripples.js
