+++ fusion/Slider.js
