+++ fusion/Truncate.js
