+++ fusion/Typing.js
