+++ index.js
@@ -1,32 +1,32 @@
-export { default as Card } from './components/Card';
-export { default as Alert } from './components/Alert';
-export { default as Button } from './components/Button';
-export { default as BackgroundImage } from './components/BackgroundImage';
-export { default as Banner } from './components/Banner';
-export { default as CountDownTimer } from './components/CountdownTimer';
-export { default as Dialog } from './components/Dialog';
-export { default as Rating } from './components/Rating';
-export { default as Panel } from './components/Panel';
-export { default as Avatar } from './components/Avatar';
-export { default as CardGrid } from './components/CardGrid';
-export { default as Navigation } from './components/Navigation';
-export { default as Heading } from './components/Heading';
-export { default as Slider } from './components/Slider';
-export { default as Typing } from './components/Typing';
-export { default as Truncate } from './components/Truncate';
-export { default as Steps } from './components/Steps';
-export { default as StepperControl } from './components/StepperControl';
-export { default as Stepper } from './components/Stepper';
-export { default as Bounce } from './components/Bounce';
-export { default as BlockQuote } from './components/Blockquote';
-export { default as Logo } from './components/Logo';
-export { default as Dropdown } from './components/Dropdown';
-export { default as Loader } from './components/Loader';
-export { default as Navbar } from './components/Navbar';
-export { default as Pagination } from './components/Pagination';
-export { default as ProgressBar } from './components/ProgressBar';
-export { default as ProgressCircle } from './components/ProgressCircle';
-export { default as Ripples } from './components/Ripples';
-export { default as Tooltip } from './components/Tooltip';
-export { default as Row } from './components/Row';
+export { default as Card } from './fusion/Card';
+export { default as Alert } from './fusion/Alert';
+export { default as Button } from './fusion/Button';
+export { default as BackgroundImage } from './fusion/BackgroundImage';
+export { default as Banner } from './fusion/Banner';
+export { default as CountDownTimer } from './fusion/CountdownTimer';
+export { default as Dialog } from './fusion/Dialog';
+export { default as Rating } from './fusion/Rating';
+export { default as Panel } from './fusion/Panel';
+export { default as Avatar } from './fusion/Avatar';
+export { default as CardGrid } from './fusion/CardGrid';
+export { default as Navigation } from './fusion/Navigation';
+export { default as Heading } from './fusion/Heading';
+export { default as Slider } from './fusion/Slider';
+export { default as Typing } from './fusion/Typing';
+export { default as Truncate } from './fusion/Truncate';
+export { default as Steps } from './fusion/Steps';
+export { default as StepperControl } from './fusion/StepperControl';
+export { default as Stepper } from './fusion/Stepper';
+export { default as Bounce } from './fusion/Bounce';
+export { default as BlockQuote } from './fusion/Blockquote';
+export { default as Logo } from './fusion/Logo';
+export { default as Dropdown } from './fusion/Dropdown';
+export { default as Loader } from './fusion/Loader';
+export { default as Navbar } from './fusion/Navbar';
+export { default as Pagination } from './fusion/Pagination';
+export { default as ProgressBar } from './fusion/ProgressBar';
+export { default as ProgressCircle } from './fusion/ProgressCircle';
+export { default as Ripples } from './fusion/Ripples';
+export { default as Tooltip } from './fusion/Tooltip';
+export { default as Row } from './fusion/Row';
 export { default as theme } from './theme/';
